Create EnergiaClienteDomain and DatabaseFunctions
Create Domain to store response objects
Create DatabaseFunctions to store internal database functions
